requirements.txt
# Image processing
Pillow>=9.5.0
numpy>=1.24.0

# Color processing and extraction
colorthief>=0.2.1
# Note: colorthief requires Pillow

# Machine learning for color clustering
scikit-learn>=1.3.0
scipy>=1.10.0

# XML/SVG processing
lxml>=4.9.0

# Async support (already in Python 3.7+, but listing for completeness)
# asyncio is built-in

# Optional: For better SVG optimization (not strictly required)
# scour>=0.38.0

# Note: This project also requires 'potrace' to be installed separately
# On macOS: brew install potrace
# On Ubuntu/Debian: sudo apt-get install potrace
# On Windows: Download from http://potrace.sourceforge.net/
